parser.rs file renamed to run.rs
